requirements.txt
Django>=1.6.1
Pillow>=2.3.0
boto>=2.26.0
dj-database-url>=0.2.2
dj-static>=0.0.5
django-analytical>=0.16.0
django-bootstrap3>=2.5.5
django-compressor>=1.3
django-heroku-memcacheify>=0.5 
django-heroku-postgresify>=0.3
django-htmlmin>=0.6.3
django-jquery>=1.9.1
django-robots>=1.0
django-storages>=1.1.8 
django-toolbelt>=0.0.1
django-pylibmc-sasl>=0.2.4
greenlet>=0.3.4
gevent>=0.13.4
gunicorn>=18.0
mandrill>=1.0.53
newrelic>=2.10.1.9
raven>=3  
sorl-thumbnail>=11.12   
psycopg2>=2.5.1
pylibmc>=1.2.3
pytz>=2013.9
static>=0.4
virtualenv>=1.11
wsgiref>=0.1.2
